Add a refresh button to PageSpeed Insights.
