+++ assets/js/modules/pagespeed-insights/components/dashboard/DashboardPageSpeed.js
@@ -31,7 +31,6 @@ import { __ } from '@wordpress/i18n';
 /**
  * Internal dependencies
  */
-import API from 'googlesitekit-api';
 import Data from 'googlesitekit-data';
 import DeviceSizeTabBar from '../../../../components/DeviceSizeTabBar';
 import ProgressBar from '../../../../components/progress-bar';
@@ -75,21 +74,11 @@ export default function DashboardPageSpeed() {
 		};
 	} );
 
-	const { invalidateResolution } = useDispatch( STORE_NAME );
 	const { setValues } = useDispatch( CORE_FORMS );
 	const setStrategyMobile = useCallback( () => setValues( FORM_DASH_WIDGET, { strategy: STRATEGY_MOBILE } ), [] );
 	const setStrategyDesktop = useCallback( () => setValues( FORM_DASH_WIDGET, { strategy: STRATEGY_DESKTOP } ), [] );
 	const setDataSrcField = useCallback( () => setValues( FORM_DASH_WIDGET, { dataSrc: DATA_SRC_FIELD } ), [] );
 	const setDataSrcLab = useCallback( () => setValues( FORM_DASH_WIDGET, { dataSrc: DATA_SRC_LAB } ), [] );
-	const updateReport = useCallback( async ( event ) => {
-		event.preventDefault();
-
-		// Invalidate the PageSpeed API request caches.
-		await API.invalidateCache( 'modules', 'pagespeed' );
-		// Invalidate the cached resolver.
-		invalidateResolution( 'getReport', [ referenceURL, STRATEGY_DESKTOP ] );
-		invalidateResolution( 'getReport', [ referenceURL, STRATEGY_MOBILE ] );
-	}, [ invalidateResolution ] );
 
 	// Update the active tab for "In the Lab" or "In The Field".
 	const updateActiveTab = useCallback( ( dataSrcIndex ) => {
@@ -137,7 +126,6 @@ export default function DashboardPageSpeed() {
 		<Fragment>
 			<header className="googlesitekit-pagespeed-widget__header">
 				<div className="googlesitekit-pagespeed-widget__data-src-tabs">
-					<a href="#update" onClick={ updateReport }>FORCE UPDATE</a>
 					<TabBar
 						activeIndex={ [ DATA_SRC_LAB, DATA_SRC_FIELD ].indexOf( dataSrc ) }
 						handleActiveIndexUpdate={ updateActiveTab }
